further changes done regarding architectural pattern MVVM
