+++ SustainU/Resources/Assets.xcassets/blueLogoColor.colorset/Contents.json
