+++ SustainU/Resources/Assets.xcassets/greenLogoColor.colorset/Contents.json
