+++ SustainU/Resources/Assets.xcassets/logoCamera.imageset/Contents.json
