+++ SustainU/Resources/Assets.xcassets/logoMap.imageset/Contents.json
