+++ SustainU/Resources/Assets.xcassets/logoRecycle.imageset/Contents.json
