+++ SustainU/Resources/Assets.xcassets/logoScoreboard.imageset/Contents.json
