+++ SustainU/Resources/Assets.xcassets/peopleCartoonLogo.imageset/Contents.json
