+++ SustainU/Resources/Assets.xcassets/redLogoColor.colorset/Contents.json
