--- SustainU/views/GridMenuView.swift
@@ -1,8 +1,0 @@
-//
-//  GridMenuView.swift
-//  SustainU
-//
-//  Created by Herrera Alba Eduardo Jose on 19/09/24.
-//
-
-import Foundation
