--- SustainU/views/HomeView.swift
@@ -1,8 +1,0 @@
-//
-//  HomeView.swift
-//  SustainU
-//
-//  Created by Herrera Alba Eduardo Jose on 19/09/24.
-//
-
-import Foundation
